refactor: move the agent-cli policy types out of the table module
policy.rs sat one line under the R8 cap after the help_search row. The
types and the classification rules move to policy/types.rs, re-exported so
every import path is unchanged; the POLICY table itself is byte-identical.
